// src/modules/Admin/interfaces/product.interface.ts
export interface createProductI {
  title: string;
  slug?: string[];
  description: string;
  price: number;
  quantity: number;
  images: string[];
  stock: number;

  categoryId: string;
  adminId: string;
}

export interface editProductI {
  title: string;
  slug?: string[];
  description: string;
  price?: number;
  quantity?: number;
  // images: string[];
  stock?: number;
  brand: any;
  colors?: string;
}